+++ vaadin-components-gwt/src/main/java/com/vaadin/client/widgets/Grid.java
@@ -168,6 +168,8 @@ import com.vaadin.shared.util.SharedUtil;
 // Fork from Vaadin 7.5.3
 // Applied changes:
 // https://dev.vaadin.com/review/#/c/11846/
+// https://dev.vaadin.com/review/#/c/11876/
+// https://dev.vaadin.com/review/#/c/11919/
 /**
  * A data grid view that supports columns and lazy loading of data rows from a
  * data source.
@@ -2787,7 +2789,9 @@ public class Grid<T> extends ResizeComposite implements
             for (Column<?, T> column : visibleColumns) {
                 final double widthAsIs = column.getWidth();
                 final boolean isFixedWidth = widthAsIs >= 0;
-                final double widthFixed = Math.max(widthAsIs,
+                // Check for max width just to be sure we don't break the limits
+                final double widthFixed = Math.max(
+                        Math.min(getMaxWidth(column), widthAsIs),
                         column.getMinimumWidth());
                 defaultExpandRatios = defaultExpandRatios
                         && (column.getExpandRatio() == -1 || column == selectionColumn);
@@ -2827,7 +2831,10 @@ public class Grid<T> extends ResizeComposite implements
             double pixelsToDistribute = escalator.getInnerWidth()
                     - reservedPixels;
             if (pixelsToDistribute <= 0 || totalRatios <= 0) {
-                setColumnSizes(columnSizes);
+                if (pixelsToDistribute <= 0) {
+                    // Set column sizes for expanding columns
+                    setColumnSizes(columnSizes);
+                }
                 return;
             }
 
@@ -7871,6 +7878,10 @@ public class Grid<T> extends ResizeComposite implements
         if (detailsGenerator == null) {
             throw new IllegalArgumentException(
                     "Details generator may not be null");
+        }
+
+        for (Integer rowIndex : visibleDetails) {
+            setDetailsVisible(rowIndex, false);
         }
 
         this.detailsGenerator = detailsGenerator;