+++ vaadin-components-gwt/src/main/java/com/vaadin/components/grid/config/JSColumn.java
@@ -18,6 +18,12 @@ public interface JSColumn {
     void setName(String s);
 
     @JsProperty
+    String getHidingToggleText();
+
+    @JsProperty
+    void setHidingToggleText(String s);
+
+    @JsProperty
     double getMinWidth();
 
     @JsProperty
@@ -58,4 +64,16 @@ public interface JSColumn {
 
     @JsProperty
     void setHeaderContent(Object o);
+
+    @JsProperty
+    boolean getHidable();
+
+    @JsProperty
+    void setHidable(boolean b);
+
+    @JsProperty
+    boolean getHidden();
+
+    @JsProperty
+    void setHidden(boolean b);
 }