+++ vaadin-components-gwt/src/main/java/com/vaadin/components/grid/table/GridLightDomTable.java
@@ -4,6 +4,7 @@ import static com.google.gwt.query.client.GQuery.$;
 
 import java.util.List;
 
+import com.google.gwt.core.client.Scheduler;
 import com.google.gwt.dom.client.Element;
 import com.google.gwt.query.client.Function;
 import com.google.gwt.query.client.GQuery;
@@ -114,6 +115,10 @@ public class GridLightDomTable implements MutationListener {
             JSColumn column = JS.createJsType(JSColumn.class);
             jsColumns.add(column);
             column.setSortable(JSValidate.Boolean.attr($th, "sortable"));
+            Scheduler.get().scheduleDeferred(
+                    () -> column.setHidable(JSValidate.Boolean.attr($th,
+                            "hidable")));
+            column.setHidden(JSValidate.Boolean.attr($th, "hidden"));
             column.setFlex(JSValidate.Integer.attr($th, "flex", 1, -1));
 
             Double width = JSValidate.Pixel.attr($th, "width");
@@ -137,6 +142,8 @@ public class GridLightDomTable implements MutationListener {
                 jsSortOrder.setColumn(i);
                 sortOrders.add(jsSortOrder);
             }
+
+            column.setHidingToggleText(JSValidate.String.attr($th, "hiding-toggle-text"));
 
             String headerHtml = JSValidate.String.attr($th, "header-text",
                     $th.html(), $th.html());
@@ -203,7 +210,7 @@ public class GridLightDomTable implements MutationListener {
                 new Timer() {
                     @Override
                     public void run() {
-                        cell.setWidget(new HTML($th.html()));
+                        js.setContent(new HTML($th.html()).getElement());
                         js.setColspan(colspan);
                     }
                 }.schedule(0);